+++ template/astro.config.mjs
@@ -2,6 +2,7 @@
 import angular from "@analogjs/astro-angular";
 import preact from "@astrojs/preact";
 import react from "@astrojs/react";
+import solid from "@astrojs/solid-js";
 import svelte from "@astrojs/svelte";
 import vue from "@astrojs/vue";
 import { defineConfig } from "astro/config";
@@ -25,6 +26,10 @@ export default defineConfig({
     }),
     react({
       include: ["src/framework/react/*"],
+    }),
+    solid({
+      devtools: true,
+      include: ["src/framework/solid/*"],
     }),
     svelte({
       include: ["src/framework/svelte/*"],