+++ template/eslint.config.mjs
@@ -12,6 +12,7 @@ import perfectionist from "eslint-plugin-perfectionist";
 import playwright from "eslint-plugin-playwright";
 import pluginReact from "eslint-plugin-react";
 import pluginReactHooks from "eslint-plugin-react-hooks";
+import solid from "eslint-plugin-solid";
 import svelte from "eslint-plugin-svelte";
 import testingLibrary from "eslint-plugin-testing-library";
 import pluginVue from "eslint-plugin-vue";
@@ -211,5 +212,9 @@ export default [
         version: "19.0",
       },
     },
+  },
+  {
+    ...solid.configs["flat/recommended"],
+    files: ["src/framework/solid/**/*.{js,ts,jsx,tsx}"],
   },
 ];