+++ template/vitest.config.ts
@@ -4,6 +4,7 @@ import preact from "@preact/preset-vite";
 import { svelte } from "@sveltejs/vite-plugin-svelte";
 import react from "@vitejs/plugin-react";
 import vue from "@vitejs/plugin-vue";
+import solid from "vite-plugin-solid";
 import { defineConfig } from "vitest/config";
 
 export default defineConfig(({ mode }) => ({
@@ -40,6 +41,17 @@ export default defineConfig(({ mode }) => ({
           globals: true,
           include: ["test/integration/react/**/*.test.tsx"],
           name: "react",
+          setupFiles: ["test/setup-test-env.ts"],
+        },
+      },
+      {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        plugins: [solid() as any],
+        test: {
+          environment: "happy-dom",
+          globals: true,
+          include: ["test/integration/solid/**/*.test.tsx"],
+          name: "solid",
           setupFiles: ["test/setup-test-env.ts"],
         },
       },